Rearranging Main.py
Rearranging
Main.py

Adding
main_scrapeminutes.py
regression.py
settings.py
visualization.py
data_processor.py
